requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
httpx==0.24.1
numpy==1.25.2
Pillow==10.0.0
sentence-transformers==2.2.2
python-dotenv==1.0.0 